requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
json-repair>=0.10.0
